scraper: parser refactoring
