+++ services/scraper/pkg/parser/table.go
@@ -1,6 +1,8 @@
 package parser
 
 import (
+	"strings"
+
 	"github.com/PuerkitoBio/goquery"
 	"github.com/arseniizyk/mgkct-schedule-bot/services/scraper/internal/models"
 )
@@ -37,4 +39,37 @@ func parseColumns(tds *goquery.Selection, days []models.Day) {
 			IsEmpty: false,
 		})
 	}
+}
+
+func parsePairs(nameParts, classParts []string) []models.Pair {
+	var pairs []models.Pair
+
+	for i := 0; i < len(nameParts); {
+		var subjectType, teacher, class string
+		name := nameParts[i]
+		i++
+
+		if i < len(nameParts) && strings.HasPrefix(nameParts[i], "(") {
+			subjectType = nameParts[i]
+			subjectType, _ = strings.CutPrefix(subjectType, "(")
+			subjectType, _ = strings.CutSuffix(subjectType, ")")
+			i++
+		}
+
+		if i < len(nameParts) {
+			teacher = nameParts[i]
+			i++
+		}
+
+		class = classParts[len(pairs)]
+
+		pairs = append(pairs, models.Pair{
+			Name:    cleanText(name),
+			Type:    cleanText(subjectType),
+			Teacher: cleanText(teacher),
+			Class:   cleanText(class),
+		})
+	}
+
+	return pairs
 }
